+++ migrate/changeset/schema.py
@@ -1,11 +1,12 @@
 """
    Schema module providing common schema operations.
 """
+from UserDict import DictMixin
 import sqlalchemy
 
+from migrate.changeset.exceptions import *
 from migrate.changeset.databases.visitor import (get_engine_visitor,
                                                  run_single_visitor)
-from migrate.changeset.exceptions import *
 
 
 __all__ = [
@@ -17,13 +18,17 @@ __all__ = [
     'ChangesetTable',
     'ChangesetColumn',
     'ChangesetIndex',
+    'ChangesetDefaultClause',
+    'ColumnDelta',
 ]
+
+DEFAULT_ALTER_METADATA = True
 
 
 def create_column(column, table=None, *p, **k):
     """Create a column, given the table
     
-    API to :meth:`column.create`
+    API to :meth:`ChangesetColumn.create`
     """
     if table is not None:
         return table.create_column(column, *p, **k)
@@ -33,7 +38,7 @@ def create_column(column, table=None, *p, **k):
 def drop_column(column, table=None, *p, **k):
     """Drop a column, given the table
     
-    API to :meth:`column.drop`
+    API to :meth:`ChangesetColumn.drop`
     """
     if table is not None:
         return table.drop_column(column, *p, **k)
@@ -45,7 +50,7 @@ def rename_table(table, name, engine=None):
 
     If Table instance is given, engine is not used.
 
-    API to :meth:`table.rename`
+    API to :meth:`ChangesetTable.rename`
 
     :param table: Table to be renamed
     :param name: new name
@@ -64,7 +69,7 @@ def rename_index(index, name, table=None, engine=None):
     If Index and Table object instances are given,
     table and engine are not used.
 
-    API to :meth:`index.rename`
+    API to :meth:`ChangesetIndex.rename`
 
     :param index: Index to be renamed
     :param name: new name
@@ -82,50 +87,25 @@ def rename_index(index, name, table=None, engine=None):
 def alter_column(*p, **k):
     """Alter a column.
 
-    Parameters: column name, table name, an engine, and the properties
-    of that column to change
+    Direct API to :class:`ColumnDelta`
 
-    API to :meth:`column.alter`
+    :param table: Table or table name (will issue reflection) 
+    :param engine: Will be used for reflection
+    :param alter_metadata: Defaults to True. It will alter changes also to objects.
     """
-    if len(p) and isinstance(p[0], sqlalchemy.Column):
-        col = p[0]
-    else:
-        col = None
+    
+    k.setdefault('alter_metadata', DEFAULT_ALTER_METADATA)
 
-    if 'table' not in k:
-        k['table'] = col.table
+    if 'table' not in k and isinstance(p[0], sqlalchemy.Column):
+        k['table'] = p[0].table
     if 'engine' not in k:
         k['engine'] = k['table'].bind
 
     engine = k['engine']
-    delta = _ColumnDelta(*p, **k)
-
-    delta.result_column.delta = delta
-    delta.result_column.table = delta.table
+    delta = ColumnDelta(*p, **k)
 
     visitorcallable = get_engine_visitor(engine, 'schemachanger')
-    engine._run_visitor(visitorcallable, delta.result_column)
-
-    # Update column
-    if col is not None:
-        # Special case: change column key on rename, if key not
-        # explicit
-        #
-        # Used by SA : table.c.[key]
-        #
-        # This fails if the key was explit AND equal to the column
-        # name.  (It changes the key name when it shouldn't.)
-        #
-        # Not much we can do about it.
-        if 'name' in delta.keys():
-            if (col.name == col.key):
-                newname = delta['name']
-                del col.table.c[col.key]
-                setattr(col, 'key', newname)
-                col.table.c[col.key] = col
-        # Change all other attrs
-        for key, val in delta.iteritems():
-            setattr(col, key, val)
+    engine._run_visitor(visitorcallable, delta)
 
 
 def _to_table(table, engine=None):
@@ -152,122 +132,250 @@ def _to_index(index, table=None, engine=None):
     return ret
 
 
-class _ColumnDelta(dict):
-    """Extracts the differences between two columns/column-parameters"""
-
-    def __init__(self, *p, **k):
-        """Extract ALTER-able differences from two columns.
+class ColumnDelta(DictMixin, sqlalchemy.schema.SchemaItem):
+    """Extracts the differences between two columns/column-parameters
 
         May receive parameters arranged in several different ways:
-         * old_column_object,new_column_object,*parameters Identifies
-            attributes that differ between the two columns.
-            Parameters specified outside of either column are always
-            executed and override column differences.
-         * column_object,[current_name,]*parameters Parameters
-            specified are changed; table name is extracted from column
-            object.  Name is changed to column_object.name from
-            current_name, if current_name is specified. If not
-            specified, name is unchanged.
-         * current_name,table,*parameters 'table' may be either an
-            object or a name
-        """
+
+        * **current_column, new_column, \*p, \*\*kw**
+            Additional parameters can be specified to override column
+            differences.
+
+        * **current_column, \*p, \*\*kw**
+            Additional parameters alter current_column. Table name is extracted
+            from current_column object.
+            Name is changed to current_column.name from current_name,
+            if current_name is specified.
+
+        * **current_col_name, \*p, \*\*kw**
+            Table kw must specified.
+
+        :param table: Table at which current Column should be bound to.\
+        If table name is given, reflection will be used.
+        :type table: string or Table instance
+        :param alter_metadata: If True, it will apply changes to metadata.
+        :type alter_metadata: bool
+        :param metadata: If `alter_metadata` is true, \
+        metadata is used to reflect table names into
+        :type metadata: :class:`MetaData` instance
+        :param engine: When reflecting tables, either engine or metadata must \
+        be specified to acquire engine object.
+        :type engine: :class:`Engine` instance
+        :returns: :class:`ColumnDelta` instance provides interface for altered attributes to \
+        `result_column` through :func:`dict` alike object.
+
+        * :class:`ColumnDelta`.result_column is altered column with new attributes
+
+        * :class:`ColumnDelta`.current_name is current name of column in db
+
+        
+    """
+
+    # Column attributes that can be altered
+    diff_keys = ('name', 'type', 'primary_key', 'nullable',
+        'server_onupdate', 'server_default')
+    diffs = dict()
+    __visit_name__ = 'column'
+
+    def __init__(self, *p, **kw):
+        self.alter_metadata = kw.pop("alter_metadata", False)
+        self.meta = kw.pop("metadata", None)
+        self.engine = kw.pop("engine", None)
+
         # Things are initialized differently depending on how many column
         # parameters are given. Figure out how many and call the appropriate
         # method.
-
         if len(p) >= 1 and isinstance(p[0], sqlalchemy.Column):
             # At least one column specified
             if len(p) >= 2 and isinstance(p[1], sqlalchemy.Column):
                 # Two columns specified
-                func = self._init_2col
+                diffs = self.compare_2_columns(*p, **kw)
             else:
                 # Exactly one column specified
-                func = self._init_1col
+                diffs = self.compare_1_column(*p, **kw)
         else:
             # Zero columns specified
-            func = self._init_0col
-        diffs = func(*p, **k)
-        self._set_diffs(diffs)
+            if not len(p) or not isinstance(p[0], basestring):
+                raise ValueError("First argument must be column name")
+            diffs = self.compare_parameters(*p, **kw)
 
-    # Column attributes that can be altered
-    diff_keys = ('name',
-                 'type',
-                 'nullable',
-                 'default',
-                 'server_default',
-                 'primary_key',
-                 'foreign_key')
+        self.apply_diffs(diffs)
 
-    @property
-    def table(self):
-        if isinstance(self._table, sqlalchemy.Table):
-            return self._table
+    def __repr__(self):
+        return '<ColumnDelta altermetadata=%r, %s>' % (self.alter_metadata,
+            super(ColumnDelta, self).__repr__())
 
-    def _init_0col(self, current_name, *p, **k):
-        p, k = self._init_normalize_params(p, k)
-        table = k.pop('table')
-        self.current_name = current_name
-        self._table = table
-        self.result_column = table.c.get(current_name, None)
+    def __getitem__(self, key):
+        if key not in self.keys():
+            raise KeyError("No such diff key, available: %s" % self.diffs )
+        return getattr(self.result_column, key)
+
+    def __setitem__(self, key, value):
+        if key not in self.keys():
+            raise KeyError("No such diff key, available: %s" % self.diffs )
+        setattr(self.result_column, key, value)
+
+    def __delitem__(self, key):
+        raise NotImplementedError
+
+    def keys(self):
+        return self.diffs.keys()
+
+    def compare_parameters(self, current_name, *p, **k):
+        """Compares Column objects with reflection"""
+        self.table = k.pop('table')
+        self.result_column = self._table.c.get(current_name)
+        if len(p):
+            k = self._extract_parameters(p, k, self.result_column)
         return k
 
-    def _init_1col(self, col, *p, **k):
-        p, k = self._init_normalize_params(p, k)
-        self._table = k.pop('table', None) or col.table
-        self.result_column = col.copy()
-        if 'current_name' in k:
-            # Renamed
-            self.current_name = k.pop('current_name')
-            k.setdefault('name', col.name)
-        else:
-            self.current_name = col.name
+    def compare_1_column(self, col, *p, **k):
+        """Compares one Column object"""
+        self.table = k.pop('table', None) or col.table
+        self.result_column = col
+        if len(p):
+            k = self._extract_parameters(p, k, self.result_column)
         return k
 
-    def _init_2col(self, start_col, end_col, *p, **k):
-        p, k = self._init_normalize_params(p, k)
-        self.result_column = start_col.copy()
-        self._table = k.pop('table', None) or start_col.table \
-            or end_col.table
-        self.current_name = start_col.name
-        for key in ('name', 'nullable', 'default', 'server_default',
-                    'primary_key', 'foreign_key'):
-            val = getattr(end_col, key, None)
-            if getattr(start_col, key, None) != val:
+    def compare_2_columns(self, old_col, new_col, *p, **k):
+        """Compares two Column objects"""
+        self.process_column(new_col)
+        self.table = k.pop('table', None) or old_col.table or new_col.table
+        self.result_column = old_col
+
+        # set differences
+        # leave out some stuff for later comp
+        for key in (set(self.diff_keys) - set(('type',))):
+            val = getattr(new_col, key, None)
+            if getattr(self.result_column, key, None) != val:
                 k.setdefault(key, val)
-        if not self.column_types_eq(start_col.type, end_col.type):
-            k.setdefault('type', end_col.type)
+
+        # inspect types
+        if not self.are_column_types_eq(self.result_column.type, new_col.type):
+            k.setdefault('type', new_col.type)
+
+        if len(p):
+            k = self._extract_parameters(p, k, self.result_column)
         return k
 
-    def _init_normalize_params(self, p, k):
-        p = list(p)
-        if len(p):
-            k.setdefault('name', p.pop(0))
-        if len(p):
-            k.setdefault('type', p.pop(0))
-        # TODO: sequences? FKs?
-        return p, k
-
-    def _set_diffs(self, diffs):
+    def apply_diffs(self, diffs):
+        """Populate dict and column object with new values"""
+        self.diffs = diffs
         for key in self.diff_keys:
             if key in diffs:
-                self[key] = diffs[key]
-                if getattr(self, 'result_column', None) is not None:
-                    setattr(self.result_column, key, diffs[key])
+                setattr(self.result_column, key, diffs[key])
 
-    def column_types_eq(self, this, that):
-        ret = isinstance(this, that.__class__)
-        ret = ret or isinstance(that, this.__class__)
+        self.process_column(self.result_column)
+
+        # create an instance of class type if not yet
+        if 'type' in diffs and callable(self.result_column.type):
+            self.result_column.type = self.result_column.type()
+
+        # add column to the table
+        if self.table and self.alter_metadata:
+            self.result_column.add_to_table(self.table)
+
+    def are_column_types_eq(self, old_type, new_type):
+        """Compares two types to be equal"""
+        ret = old_type.__class__ == new_type.__class__
+
         # String length is a special case
-        if ret and isinstance(that, sqlalchemy.types.String):
-            ret = (getattr(this, 'length', None) == \
-                       getattr(that, 'length', None))
+        if ret and isinstance(new_type, sqlalchemy.types.String):
+            ret = (getattr(old_type, 'length', None) == \
+                       getattr(new_type, 'length', None))
         return ret
+
+    def _extract_parameters(self, p, k, column):
+        """Extracts data from p and modifies diffs"""
+        p = list(p)
+        while len(p):
+            if isinstance(p[0], basestring):
+                k.setdefault('name', p.pop(0))
+            elif isinstance(p[0], sqlalchemy.types.AbstractType):
+                k.setdefault('type', p.pop(0))
+            elif callable(p[0]):
+                p[0] = p[0]()
+            else:
+                break
+
+        if len(p):
+            new_col = column.copy_fixed()
+            new_col._init_items(*p)
+            k = self.compare_2_columns(column, new_col, **k)
+        return k
+
+    def process_column(self, column):
+        """Processes default values for column"""
+        # XXX: this is a snippet from SA processing of positional parameters
+        if column.args:
+            toinit = list(column.args)
+        else:
+            toinit = list()
+
+        if column.server_default is not None:
+            if isinstance(column.server_default, sqlalchemy.FetchedValue):
+                toinit.append(column.server_default)
+            else:
+                toinit.append(sqlalchemy.DefaultClause(column.server_default))
+        if column.server_onupdate is not None:
+            if isinstance(column.server_onupdate, FetchedValue):
+                toinit.append(column.server_default)
+            else:
+                toinit.append(sqlalchemy.DefaultClause(column.server_onupdate,
+                                            for_update=True))
+        if toinit:
+            column._init_items(*toinit)
+        column.args = []
+
+    def _get_table(self):
+        return getattr(self, '_table', None)
+
+    def _set_table(self, table):
+        if isinstance(table, basestring):
+            if self.alter_metadata:
+                if not self.meta:
+                    raise ValueError("metadata must be specified for table"
+                        " reflection when using alter_metadata")
+                meta = self.meta
+                if self.engine:
+                    meta.bind = self.engine
+            else:
+                if not self.engine and not self.meta:
+                    raise ValueError("engine or metadata must be specified"
+                        " to reflect tables")
+                if not self.engine:
+                    self.engine = self.meta.bind
+                meta = sqlalchemy.MetaData(bind=self.engine)
+            self._table = sqlalchemy.Table(table, meta, autoload=True)
+        elif isinstance(table, sqlalchemy.Table):
+            self._table = table
+            if not self.alter_metadata:
+                self._table.meta = sqlalchemy.MetaData(bind=self._table.bind)
+
+    def _get_result_column(self):
+        return getattr(self, '_result_column', None)
+
+    def _set_result_column(self, column):
+        """Set Column to Table based on alter_metadata evaluation."""
+        self.process_column(column)
+        if not hasattr(self, 'current_name'):
+            self.current_name = column.name
+        if self.alter_metadata:
+            self._result_column = column
+            # remove column from table, nothing has changed yet
+            if self.table:
+                column.remove_from_table(self.table)
+        else:
+            self._result_column = column.copy_fixed()
+
+    table = property(_get_table, _set_table)
+    result_column = property(_get_result_column, _set_result_column)
 
 
 class ChangesetTable(object):
     """Changeset extensions to SQLAlchemy tables."""
 
-    def create_column(self, column):
+    def create_column(self, column, **kw):
         """Creates a column.
 
         The column parameter may be a column definition or the name of
@@ -278,7 +386,7 @@ class ChangesetTable(object):
             column = getattr(self.c, str(column))
         column.create(table=self)
 
-    def drop_column(self, column):
+    def drop_column(self, column, **kw):
         """Drop a column, given its name or definition."""
         if not isinstance(column, sqlalchemy.Column):
             # It's a column name
@@ -327,17 +435,16 @@ class ChangesetColumn(object):
         May supply a new column object, or a list of properties to
         change.
 
-        For example; the following are equivalent:
-            col.alter(Column('myint', Integer, nullable=False))
-            col.alter('myint', Integer, nullable=False)
-            col.alter(name='myint', type=Integer, nullable=False)
+        For example; the following are equivalent::
 
-        Column name, type, default, and nullable may be changed
-        here. Note that for column defaults, only PassiveDefaults are
-        managed by the database - changing others doesn't make sense.
+            col.alter(Column('myint', Integer, DefaultClause('foobar')))
+            col.alter('myint', Integer, server_default='foobar', nullable=False)
+            col.alter(DefaultClause('foobar'), name='myint', type=Integer, nullable=False)
 
-        :param table: Table to be altered
-        :param engine: Engine to be used
+        Column name, type, server_default, and nullable may be changed
+        here.
+
+        Direct API to :func:`alter_column`
         """
         if 'table' not in k:
             k['table'] = self.table
@@ -371,8 +478,8 @@ class ChangesetColumn(object):
         """
         if table is not None:
             self.table = table
-        self.remove_from_table(self.table)
         engine = self.table.bind
+        self.remove_from_table(self.table, unset_table=False)
         visitorcallable = get_engine_visitor(engine, 'columndropper')
         engine._run_visitor(visitorcallable, self, *args, **kwargs)
         return self
@@ -381,12 +488,31 @@ class ChangesetColumn(object):
         if table and not self.table:
             self._set_parent(table)
 
-    def remove_from_table(self, table):
+    def remove_from_table(self, table, unset_table=True):
         # TODO: remove indexes, primary keys, constraints, etc
+        if unset_table:
+            self.table = None
         if table.c.contains_column(self):
             table.c.remove(self)
             
+    # TODO: this is fixed in 0.6
+    def copy_fixed(self, **kw):
+        """Create a copy of this ``Column``, with all attributes."""
+        return sqlalchemy.Column(self.name, self.type, self.default,
+            key=self.key,
+            primary_key=self.primary_key,
+            nullable=self.nullable,
+            quote=self.quote,
+            index=self.index,
+            unique=self.unique,
+            onupdate=self.onupdate,
+            autoincrement=self.autoincrement,
+            server_default=self.server_default,
+            server_onupdate=self.server_onupdate,
+            *[c.copy(**kw) for c in self.constraints])
+
     def _check_sanity_constraints(self, name):
+
         obj = getattr(self, name)
         if (getattr(self, name[:-5]) and not obj):
             raise InvalidConstraintError("Column.create() accepts index_name,"
@@ -412,3 +538,15 @@ class ChangesetIndex(object):
         visitorcallable = get_engine_visitor(engine, 'schemachanger')
         engine._run_visitor(visitorcallable, self, *args, **kwargs)
         self.name = name
+
+
+class ChangesetDefaultClause(object):
+    """Implements comparison between :class:`DefaultClause` instances"""
+
+    def __eq__(self, other):
+        if isinstance(other, self.__class__):
+            if self.arg == other.arg:
+                return True
+
+    def __ne__(self, other):
+        return not self.__eq__(other)